requirements.txt
torch
transformers
peft
accelerate
bitsandbytes
datasets
fastapi
uvicorn[standard]
pydantic
safetensors
tqdm
huggingface_hub
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart